requirements.txt
flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
flasgger==0.9.7.1
Flask-Mail==0.9.1
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1